feat: summarization improvements
- add loading state for summarizer
- add summary generation is on-the-fly and cached, rather than generated
  by AI at the point of article fetching
- refactor out article display components
