+++ components/ChatStuff/ArticleConvo.tsx
@@ -211,8 +211,8 @@ export default function ArticleConvo({ textFromArticle, chatError }: IArticleCon
 
 
         //Send message to OpenAI to get response
-        //const response = await generatePrompts('gpt-3.5-turbo', userMessage, finterestGenerateArticlePrompt.finterestGenerateArticlePrompt + currArticle.content, previousMessages);
-        const response = "Sample response 1";
+        const response = await generatePrompts('gpt-3.5-turbo', userMessage, finterestGenerateArticlePrompt.finterestGenerateArticlePrompt + currArticle.content, previousMessages);
+        // const response = "Sample response 1";
 
         console.log("Response" + response);
