+++ components/common/LoadingSpinner.tsx
@@ -1,0 +1,9 @@
+export const LoadingSpinner = () => {
+    return (
+        <div className="inline-block h-6 w-6 animate-spin rounded-full border-4 border-solid
+        border-current border-r-transparent align-[-0.125em] motion-reduce:animate-[spin_1.5s_linear_infinite]"
+            role="status">
+            <span className="!absolute !-m-px !h-px !w-px !overflow-hidden !whitespace-nowrap !border-0 !p-0 ![clip:rect(0,0,0,0)]">Loading...</span>
+        </div>
+    )
+}
