+++ lib/SummaryController.tsx
@@ -1,0 +1,24 @@
+import { updateArticleSummary } from "@/config/firestore";
+import { generatePrompts } from "@/utils/openai";
+
+
+export default async function summarizeArticleAndStoreInDb(articleId: string, articleContent: string): Promise<string> {
+    let summary = await summarizeArticleWWithAI(articleContent);
+
+    if (!summary) {
+        return "Unable to return summary."
+    } else {
+        updateArticleSummary(articleId, summary);
+        return summary;
+    }
+}
+
+async function summarizeArticleWWithAI(articleContent: string): Promise<string | null> {
+    const promptString = `Please provide a bite-sized summary of the following article in 3-8 sentences or less, relative to the initial article length. Ensure that the summary is concise, and reads like a narrative.`;
+
+    console.log("generating summary for " + articleContent)
+    const response = await generatePrompts('gpt-3.5-turbo', articleContent, promptString);
+    console.log("summary generated by openai " + response);
+    return response;
+
+}
